feat: 연관관계 메서드작성
